pyproject.toml
[project]
name = "culinary-canvas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "daphne>=4.2.1",
    "django>=5.2.6",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.9.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "redis>=6.4.0",
]